Moving pad_arrays back to mx.batchify since it uses mxnet
