+++ src/gluonts/mx/batchify.py
@@ -11,14 +11,15 @@
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
 
-from typing import List, Optional
+from typing import List, Optional, Union
+import functools
 
 import numpy as np
 import mxnet as mx
 
 from gluonts.core.component import DType
 from gluonts.dataset.common import DataBatch
-from gluonts.support.util import pad_arrays
+from gluonts.support.util import pad_to_size
 
 
 # TODO: should the following contempate mxnet arrays or just numpy arrays?
@@ -33,6 +34,31 @@ def _is_stackable(arrays: List, axis: int = 0) -> bool:
     return True
 
 
+# TODO: should the following contempate mxnet arrays or just numpy arrays?
+def _pad_arrays(
+    data: List[Union[np.ndarray, mx.nd.NDArray]], axis: int = 0
+) -> List[Union[np.ndarray, mx.nd.NDArray]]:
+    assert isinstance(data[0], (np.ndarray, mx.nd.NDArray))
+    is_mx = isinstance(data[0], mx.nd.NDArray)
+
+    # MxNet causes a segfault when persisting 0-length arrays. As such,
+    # we add a dummy pad of length 1 to 0-length dims.
+    max_len = max(1, functools.reduce(max, (x.shape[axis] for x in data)))
+    padded_data = []
+
+    for x in data:
+        # MxNet lacks the functionality to pad n-D arrays consistently.
+        # We fall back to numpy if x is an mx.nd.NDArray.
+        if is_mx:
+            x = x.asnumpy()
+
+        x = pad_to_size(x, max_len, axis)
+
+        padded_data.append(x if not is_mx else mx.nd.array(x))
+
+    return padded_data
+
+
 def stack(
     data,
     ctx: Optional[mx.context.Context] = None,
@@ -40,7 +66,7 @@ def stack(
     variable_length: bool = False,
 ):
     if variable_length and not _is_stackable(data):
-        data = pad_arrays(data, axis=0)
+        data = _pad_arrays(data, axis=0)
     if isinstance(data[0], mx.nd.NDArray):
         # TODO: think about using shared context NDArrays
         #  https://github.com/awslabs/gluon-ts/blob/42bee73409f801e7bca73245ca21cd877891437c/src/gluonts/dataset/parallelized_loader.py#L157
